requirements.txt
asgiref==3.8.1
Django==5.2
django-environ==0.12.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
mysqlclient==2.2.7
promise==2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
setuptools==58.1.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2